% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oregon7Q10.R
\name{Oregon7Q10}
\alias{Oregon7Q10}
\title{Oregon 7Q10}
\usage{
Oregon7Q10(Station_ID, start_date, end_date, period = "Annual",
  custom_start = NA, custom_end = NA, wy_start = "10-01",
  method_type = "Both")
}
\arguments{
\item{Station_ID}{Oregon Water Resources Divison (http://apps.wrd.state.or.us/apps/sw/hydro_near_real_time/) Station ID number}

\item{start_date}{Start date, in the format of "mm/dd/yyyy", for the period of calculation}

\item{end_date}{End date, in the format of "mm/dd/yyyy", for the period of calculation}

\item{period}{The temporal scale for the 7Q10 calculation, can be "Annual" or "Custom". Default is "Annual"}

\item{custom_start}{If "Custom" is designated as the period, the start date (excluding year) of the custom period in the character format of "mm-dd"}

\item{custom_end}{If "Custom" is designated as the period, the end date (excluding year) of the custom period in the character format of "mm-dd"}

\item{wy_start}{Only applicable for the "Annual" period. The water year beginning date (excluding year). The date should be in the character format of "mm-dd". If not specified the default is "10-01"}

\item{method_type}{The method used to calculate 7Q10.  Options are "Distribution Free", "log Pearson Type III", or "Both"}
}
\description{
This package contains a function returns the lowest seven-day average discharge expected once every 10 years
based on a continuous record of data.  The 7Q10 can reflect an annual, seasonal, or monthly
statistic. Methods are consistent with the EPA's DFLOW methods.  Also included is a function,
OWRD_stations, that links you to the Oregon Water Resources website to browse station IDs.
References:
https://www.epa.gov/waterdata/dflow
https://nepis.epa.gov/Exe/ZyPDF.cgi/30001JEH.PDF?Dockey=30001JEH.PD
http://deq1.bse.vt.edu/sifnwiki/index.php/Multiple_ways_to_calculate_7Q10
https://github.com/rmichie/DFLOW_R
}
\examples{
Oregon7Q10(Station_ID, start_date, end_date, period = "Annual", custom_start = NA, custom_end = NA, wy_start = "10-01", method_type = "Both")
}
\keyword{7Q10}
\keyword{Flow}
\keyword{Oregon}
